requirements.txt
sqlalchemy==1.3.20
psycopg2-binary==2.8.6
request==2.25.1
